git-tag: Do not assume the working tree root is writable.
This is a long overdue companion commit that fixed git-commit
(Santi's 6afbeb7e2714dbdc61351fe7decd1584e24bdeb6).

Having the temporary files in the working tree root when making
tags is not as bad because it does not involve 'git status' as
the git-commit case, but this makes things more consistent.

Signed-off-by: Junio C Hamano <junkio@cox.net>